refactor(meta): flatten facebook/ subpackage into root integrations
- Merge client.py + exceptions.py + models.py + utils.py into _fi_meta_helpers.py
- Inline campaigns/adsets/ads/audiences/pixels/targeting/rules into fi_meta_marketing_manage.py
- Inline insights.py into fi_meta_marketing_metrics.py
- Inline accounts.py into fi_meta_pages.py
- Delete facebook/ subpackage entirely

All fi_meta_*.py files now self-contained with single helper dep.
